fix: 디테일 동작 에러 수정 및 파일 경로 수정
