+++ src/app/auth/@signup/page.tsx
@@ -1,6 +1,5 @@
+import SignupForm from "@/components/pages/auth/@signup/SignupForm"
 import { HydrationBoundary } from "@tanstack/react-query"
-
-import SignupForm from "./SignupForm"
 
 const SignupPage = () => {
   return (
