+++ src/components/public/gnb/GNB.tsx
@@ -3,9 +3,9 @@
 import Link from "next/link"
 import { usePathname, useSearchParams } from "next/navigation"
 
-import { useGetUserData } from "@/actions/api-hooks/Auths"
 import Logo from "@/components/public/img/Logo"
 import ROUTE from "@/constants/route"
+import useGetUserData from "@/hooks/useGetUserData"
 
 import ProfileComponent from "./components/ProfileComponent"
